Extract Schema::Check definition into a separate file
