+++ lib/dry/validation/schema/check.rb
@@ -1,0 +1,20 @@
+module Dry
+  module Validation
+    class Schema
+      class Check < Value
+        private
+
+        def method_missing(meth, *meth_args)
+          vals, args = meth_args.partition { |arg| arg.class < DSL }
+
+          keys = [name, *vals.map(&:name)]
+          predicate = [:predicate, [meth, args]]
+
+          rule = create_rule([:check, [name, predicate, keys]])
+          add_rule(rule)
+          rule
+        end
+      end
+    end
+  end
+end
