+++ lib/dry/validation/schema/value.rb
@@ -6,21 +6,6 @@ module Dry
       class Value < DSL
         attr_reader :type
 
-        class Check < Value
-          private
-
-          def method_missing(meth, *meth_args)
-            vals, args = meth_args.partition { |arg| arg.class < DSL }
-
-            keys = [name, *vals.map(&:name)]
-            predicate = [:predicate, [meth, args]]
-
-            rule = create_rule([:check, [name, predicate, keys]])
-            add_rule(rule)
-            rule
-          end
-        end
-
         def initialize(options = {})
           super
           @type = options.fetch(:type, :key)
